ml_training/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
tqdm==4.66.1
onnx==1.15.0
coremltools==7.1
